apps/smart-home-ai-agent/pyproject.toml
[project]
name = "smart-home-ai-agent"
version = "0.1.0"
description = "app"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "certifi>=2024.12.14",
  "fastapi>=0.115.6",
  "japanize-matplotlib>=1.1.3",
  "langchain>=0.3.12",
  "langchain-aws>=0.2.9",
  "langchain-community>=0.3.12",
  "langgraph>=0.2.59",
  "matplotlib>=3.10.0",
  "mcp>=1.1.2",
  "pandas>=2.2.3",
  "pytz>=2024.2",
  "slack-sdk>=3.34.0",
]

[dependency-groups]
dev = [
  "freezegun>=1.5.1",
  "openai>=1.58.1",
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.0",
  "pytest-sugar>=1.0.0",
  "python-dotenv>=1.0.1",
  "ruff>=0.8.2",
]

[tool.pytest.ini_options]
env_files = [".env"]
